CLI for JUPYT
